api/requirements.txt
flask
flask_sqlalchemy==3.1.1
mysqlclient==2.2.5
Flask-Migrate==4.0.7
PyMySQL==1.1.1
cryptography==43.0.3